提交一波

+++ code/myway-RBAC/src/main/java/com/myway/rbac/controller/UserController.java
@@ -19,7 +19,7 @@ public class UserController {
     UserService userService;
 
     @RequestMapping("/get")
-    public UserDO get(Long id) {
+    public UserDO get(Integer id) {
         UserDO user = userService.get(id);
         return user;
     }
